Declare new locals in target method instead of capturing locals in patcher method

+++ RiskOfChaos/EffectDefinitions/World/AllCostsHealth.cs
@@ -139,7 +139,7 @@ namespace RiskOfChaos.EffectDefinitions.World
             return Mathf.Max(1f, (1f - (halfwayValue / (cost + halfwayValue))) * 100f);
         }
 
-        void onCharacterMoneyChanged(CharacterMaster master, int moneyDiff)
+        void onCharacterMoneyChanged(CharacterMaster master, long moneyDiff)
         {
             if (moneyDiff <= 0 || !master.playerCharacterMasterController)
                 return;
